Proof of concept release
when installing this extension you can now add media variations to an image in the TYPO3 backend. This could be the same image, the same image cropped to a different view, or a completely different (cropped) image.
A fluid-partial to replace the standard image partial to create the picture tag is provided as well.
Documentation and Integration-features will follow.