[TASK] refactor saving in reference
As the backend only saves one cropping information by default, it is
necessary to add cropping hidden input for every file reference.
Additionally, a DataHandler hook manipulates and merges the cropping
information to be saved in the original reference, so the crop can be
used in frontend directly instead of getting information from the
variant.